+++ ft_strlcpy.c
@@ -3,28 +3,28 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcpy.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: snagat <marvin@42.fr>                      +#+  +:+       +#+        */
+/*   By: snagat <snagat@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/05 14:56:50 by snagat            #+#    #+#             */
-/*   Updated: 2021/11/05 14:56:53 by snagat           ###   ########.fr       */
+/*   Updated: 2021/11/18 20:31:06 by snagat           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned        int     ft_strlcpy(char *dest, char *src, unsigned int size)
+size_t	ft_strlcpy(char *dest, char *src, unsigned int size)
 {
-        unsigned int    i;
+	unsigned int	i;
 
-        i = 0;
-        if (size > 0)
-        {
-                while (src[i] != '\0' && i < (size - 1))
-                {
-                        dest[i] = src[i];
-                        i++;
-                }
-                dest[i] = '\0';
-        }
-        return (ft_strlen(src));
+	i = 0;
+	if (size > 0)
+	{
+		while (src[i] != '\0' && i < (size - 1))
+		{
+			dest[i] = src[i];
+			i++;
+		}
+		dest[i] = '\0';
+	}
+	return (ft_strlen(src));
 }
